+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 ## [0.0.1] - 4/15/21
 
 * Added a simple notification wall that handle the initialization of Firebase, background and Local notifications.
+
+## [0.0.9] - 4/15/21
+
+* Added setting up callback widget
+
+## [0.0.10] - 4/15/21
+
+* Added onTokenRefresh callback